1) app_utils.P: modify processBodyLatest/5 to deal with 'prolog' goals (now use quintuple like other goals, rahter than anonymous var).
2) preprocess.P:
modify loadBody/2 to deal with 'prolog' goals (now preprocessed 'prolog' goals are triple) so that they can be handled uniformly in dualizing as with other (non-prolog) goals.

3) qualm.P:
add a clause for not_overridden/6 to deal with 'prolog' goals (to avoid error in computing atom_compl/2 as done in the other not_overridden/6 clause.

4) add a queries scenario q_th1a for th1a.ab